feat: added new fields and complety refactoring
