+++ src/pages/offersManager/OffersManager.css
